pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yellhorn-mcp"
version = "0.6.0"
authors = [{ name = "Author" }]
description = "Yellhorn offers MCP tools to generate detailed workplans with Gemini 2.5 Pro or OpenAI models and to review diffs against them using your entire codebase as context."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "mcp[cli]~=1.10.1",
    "google-genai~=1.24.0",
    "aiohttp~=3.12.13",
    "pydantic~=2.11.7",
    "openai~=1.93.0",
    "jedi~=0.19.2",
]

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
    "isort",
    "pytest",
    "pytest-asyncio",
    "httpx",
    "pytest-cov",
    "jedi~=0.19",
]

[project.scripts]
yellhorn-mcp = "yellhorn_mcp.cli:main"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.flake8]
max-line-length = 100
exclude = ["venv", ".git", "__pycache__", "build", "dist"]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"

[tool.hatch.build.targets.wheel]
packages = ["yellhorn_mcp"]
